+++ city_metrix/layers/overture_buildings_dsm.py
@@ -49,13 +49,18 @@ class OvertureBuildingsDSM(Layer):
             polygon = building['geometry']
 
             # Rasterize footprint to create a mask
-            building_mask = geometry_mask([polygon], transform=DEM.rio.transform(), invert=True, out_shape=DEM.shape)
+            building_mask = geometry_mask([polygon], transform=DEM.rio.transform(), all_touched=True, out_shape=DEM.shape)
 
-            # mean elevation within footprint
-            footprint_elevation = DEM.values[building_mask].mean()
+            # Get aggregated elevation within footprint
+            # TODO Implement a more sophisticated method. See https://gfw.atlassian.net/browse/CDB-309
+            DEM_values = DEM.values[building_mask]
+            # Only include features that arg large enough to overlap or touch a raster cell
+            # This situation should not happen, so the check is mostly precautionary
+            if DEM_values.size > 0:
+                footprint_elevation = DEM_values.max()
 
-            # Add the building height and mean elevation
-            DEM_updated.values[building_mask] = footprint_elevation + height
+                # Add the building height and mean elevation
+                DEM_updated.values[building_mask] = footprint_elevation + height
 
         # Clip back to the original AOI
         west, south, east, north = bbox.as_utm_bbox().bounds